introduce child aware model tranformer
Using a form on a document with an image assigned as a child, we have to set the node name which is the destination on the parent node. So this transformer does this job by passing im an option.

use "empty_data" of parent form type to get the existing file, do not set name for children

test the transformer

add functional persistence tests

separated into two transformers

revert rename in originally file

patch from Style CI

remove old method